+++ src/__tests__/Filter/noneSelected.test.tsx
@@ -7,7 +7,7 @@ jest.mock("next/navigation", () => ({
   }),
 }));
 
-import Filter from "@/components/ui/Filter";
+import Filter from "@/app/Filter";
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 import { labels } from "../fixtureData/labels";
