+++ src/app/buttons/FilterByLabelButton.tsx
@@ -1,7 +1,8 @@
 import * as React from "react";
 
 import { cn } from "@/lib/utils";
-import { Button } from "./button";
+import { Button } from "@/components/ui/button";
+
 
 function FilterByLabelButton(
   props: React.ComponentProps<"button"> & {
